LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-transition-default-expected.txt

Anchor1
Anchor2

FAIL Transition when position-anchor changes assert_equals: expected 25 but got 50
